+++ tests/test_validation.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 import pytest
 
-from security_id.validation import Cusip, Isin, Sedol
+from security_id.validation import Cusip, Isin, Sedol, val_check_digit
 from security_id.exceptions import (CharacterError, CheckDigitError, CheckSumError,
                                     CountryCodeError, LengthError, NullError)
 
@@ -136,3 +136,11 @@ def test_cusip_country_code_error():
     with pytest.raises(CountryCodeError):
         Cusip().validate('I0303M109')
 
+# --------------------------------------------------------------
+# helper tests
+# --------------------------------------------------------------
+
+class TestValCheckDigit(TestCase):
+    def test_val_check_digit(self):
+        self.assertEqual(val_check_digit('abc1'),1)
+        self.assertRaises(CheckDigitError,val_check_digit,'abc')
